@@ -13,12 +13,14 @@ dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.0
-  equatable: ^1.2.5
-  flutter_bloc: ^6.1.1
+  cupertino_icons: ^1.0.2
+  equatable: ^2.0.0
+  flutter_bloc: ^6.1.3
   dio: ^3.0.10
-  google_maps_flutter: ^1.1.1
-  geolocator: ^6.1.14
+  google_maps_flutter: ^2.0.1
+  geolocator: ^7.0.1
+  carousel_slider: ^3.0.0
+  flutter_screenutil: ^4.0.3+3
 
 dev_dependencies:
   flutter_test:
